+++ core/type/src/index.ts
@@ -1,8 +1,8 @@
+export * from './chat.js';
+export * from './service-response.js';
 export * from './storage.js';
-export * from './service.js';
 export * from './alwatr.js';
 export * from './math.js';
-export * from './chat.js';
 
 // FIXME:
 export type MaybePromise<T> = T | Promise<T>