+++ services/comment/src/lib/storage.ts
@@ -2,7 +2,7 @@ import {AlwatrStorageClient} from '@alwatr/storage-client';
 
 import {config} from '../config.js';
 
-import type {Message} from '@alwatr/type';
+import type {ChatMessage} from '@alwatr/type';
 
 
-export const storageClient = new AlwatrStorageClient<Message>(config.storage);
+export const storageClient = new AlwatrStorageClient<ChatMessage>(config.storage);