+++ ui/ui-kit/src/chat/chat-bubble.ts
@@ -17,24 +17,28 @@ export class AlwatrChatBubble extends DirectionMixin(AlwatrDummyElement) {
   static override styles = css`
     :host {
       display: inline-block;
-      padding: var(--md-sys-spacing-track-1) var(--md-sys-spacing-track-2) var(--md-sys-spacing-halftrack-3);
-      color: var(--md-sys-color-on-surface);
-      background-color: var(--md-sys-color-surface);
+      padding:
+        var(--alwatr-sys-spacing-track-1)
+        var(--alwatr-sys-spacing-track-2)
+        var(--alwatr-sys-spacing-halftrack-3)
+      ;
+      color: var(--alwatr-sys-color-on-surface);
+      background-color: var(--alwatr-sys-color-surface);
       white-space: pre-line;
-      /* max-width: var(--md-sys-spacing-column-3); */
-      border-radius: var(--md-sys-shape-large);
+      /* max-width: var(--alwatr-sys-spacing-column-3); */
+      border-radius: var(--alwatr-sys-shape-large);
     }
 
     :host([side='start']),
     :host([side='end'][dir='rtl']) {
-      border-bottom-left-radius: var(--md-sys-shape-corner-extra-small-default-size);
-      border-bottom-right-radius: var(--md-sys-shape-large);
+      border-bottom-left-radius: var(--alwatr-sys-shape-corner-extra-small-default-size);
+      border-bottom-right-radius: var(--alwatr-sys-shape-large);
     }
 
     :host([side='end']),
     :host([side='start'][dir='rtl']) {
-      border-bottom-left-radius: var(--md-sys-shape-large);
-      border-bottom-right-radius: var(--md-sys-shape-corner-extra-small-default-size);
+      border-bottom-left-radius: var(--alwatr-sys-shape-large);
+      border-bottom-right-radius: var(--alwatr-sys-shape-corner-extra-small-default-size);
     }
   `;