+++ ui/element/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.27.0](https://github.com/AliMD/alwatr/compare/v0.26.0...v0.27.0) (2022-12-29)
+
+### Bug Fixes
+
+- **element:** build issues ([0bfb016](https://github.com/AliMD/alwatr/commit/0bfb0167a82c1dfbd04884a049b76cf7f6033f86))
+- **element:** signal mixing safety check ([2f3c871](https://github.com/AliMD/alwatr/commit/2f3c87141e162e8f600246bf863fcfe93f41f53b))
+- **ui/element:** listener list type ([e089d01](https://github.com/AliMD/alwatr/commit/e089d018ed29eae8f0ba92767b12d37f5ad1729c))
+
+### Features
+
+- **element/root:** host box-sizing ([0c4ce83](https://github.com/AliMD/alwatr/commit/0c4ce83c5f777aecb2317c832817854ff4cadcf9))
+- **element:** AlwatrRootElement ([3b4e59f](https://github.com/AliMD/alwatr/commit/3b4e59f840f600ebedf07c27a9e4cf4c5f1253c0))
+- **element:** DirectionMixin ([d2670dd](https://github.com/AliMD/alwatr/commit/d2670dd247f93759e3b6457b339a066cc2433269))
+- **element:** export from lit ([21f15d8](https://github.com/AliMD/alwatr/commit/21f15d86319a40fab02d86f24a1451c056f36a7c))
+- **element:** LocalizeMixin ([94f0f7f](https://github.com/AliMD/alwatr/commit/94f0f7f832ff885d7176884f243c57a2a03c7bdb))
+- **element:** use logger ([95708a7](https://github.com/AliMD/alwatr/commit/95708a7b8a8418ffe0cb9a98c6d5ed50ff6308e1))
+- **page-chat:** new demo ([451d63f](https://github.com/AliMD/alwatr/commit/451d63fc0b904bb780c898c2909b1d7a116b353f))
+- **ui/element/i18n:** localization mixin ([999e1b1](https://github.com/AliMD/alwatr/commit/999e1b11f0af55704fdd5859398908aeedbb833c))
+- **ui/element:** seprate mixins ([a305c5d](https://github.com/AliMD/alwatr/commit/a305c5d0e027b4c19fb1a22e4dc75797d205655a))
+
 # [0.26.0](https://github.com/AliMD/alwatr/compare/v0.25.0...v0.26.0) (2022-12-22)
 
 ### Bug Fixes